# internal/api/docs/swagger.yaml
definitions:
...
  entity.Purchase:
    properties:
      description:
        type: string
      items:
        items:
          $ref: '#/definitions/entity.PurchaseItem'
        type: array
      payment_method:
        type: string
      supplier_id:
        type: string
    type: object
  entity.PurchaseItem:
    properties:
      product_id:
        type: string
      purchase_price:
        type: number
      quantity:
        type: integer
    type: object
  entity.PurchaseUpdate:
    properties:
      description:
        type: string
      id:
        type: string
      payment_method:
        type: string
      supplier_id:
        type: string
    type: object
...
  products.Error:
    properties:
      error:
        type: string
    type: object
...
  products.Message:
    properties:
      message:
        type: string
    type: object
...
  products.PurchaseItemResponse:
    properties:
      id:
        type: string
      product_id:
        type: string
      product_image:
        type: string
      product_name:
        type: string
      purchase_id:
        type: string
      purchase_price:
        description: Changed to double
        type: number
      quantity:
        type: integer
      total_price:
        description: Changed to double
        type: number
    type: object
  products.PurchaseList:
    properties:
      purchases:
        items:
          $ref: '#/definitions/products.PurchaseResponse'
        type: array
    type: object
  products.PurchaseResponse:
    properties:
      branch_id:
        description: Added branch_id
        type: string
      company_id:
        description: Company ID added
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/products.PurchaseItemResponse'
        type: array
      payment_method:
        type: string
      purchased_by:
        type: string
      purchaser_phone_number:
        type: string
      supplier_id:
        type: string
      supplier_name:
        type: string
      total_cost:
        description: Changed to double
        type: number
    type: object
...
info:
  contact: {}
  description: This is a sample server
  title: API Gateway
  version: "1.0"
paths:
...
  /purchases:
    get:
      consumes:
      - application/json
      description: Retrieve a list of purchases with optional filters
      parameters:
      - in: query
        name: created_at
        type: string
      - in: query
        name: product_id
        type: string
      - in: query
        name: purchased_by
        type: string
      - in: query
        name: supplier_id
        type: string
      - description: Branch ID
        in: header
        name: branch_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/products.PurchaseList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/products.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/products.Error'
      security:
      - ApiKeyAuth: []
      summary: List all purchases
      tags:
      - Purchases
    post:
      consumes:
      - application/json
      description: Create a new purchase with the provided details
      parameters:
      - description: Purchase data
        in: body
        name: Purchase
        required: true
        schema:
          $ref: '#/definitions/entity.Purchase'
      - description: Branch ID
        in: header
        name: branch_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/products.PurchaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/products.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/products.Error'
      security:
      - ApiKeyAuth: []
      summary: Create a new purchase
      tags:
      - Purchases
  /purchases/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a purchase by ID
      parameters:
      - description: Purchase ID
        in: path
        name: id
        required: true
        type: string
      - description: Branch ID
        in: header
        name: branch_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/products.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/products.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/products.Error'
      security:
      - ApiKeyAuth: []
      summary: Delete a purchase
      tags:
      - Purchases
    get:
      consumes:
      - application/json
      description: Retrieve a purchase by ID
      parameters:
      - description: Purchase ID
        in: path
        name: id
        required: true
        type: string
      - description: Branch ID
        in: header
        name: branch_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/products.PurchaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/products.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/products.Error'
      security:
      - ApiKeyAuth: []
      summary: Get a purchase
      tags:
      - Purchases
    put:
      consumes:
      - application/json
      description: Update the details of an existing purchase by ID
      parameters:
      - description: Purchase ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated purchase data
        in: body
        name: Purchase
        required: true
        schema:
          $ref: '#/definitions/entity.PurchaseUpdate'
      - description: Branch ID
        in: header
        name: branch_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/products.PurchaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/products.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/products.Error'
      security:
      - ApiKeyAuth: []
      summary: Update an existing purchase
      tags:
      - Purchases
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"